break up large index.js with utils.js
